+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ## 5.0.0-nullsafety.0
 
 * Null Safety support
+* [breaking change] TitledNavigationBarItem.icon is now of type Widget rather than IconData.
+* Added TitledBottomNavigationBar.height
+* Added TitledBottomNavigationBar.indicatorHeight
 
 ## 4.1.0